+++ cli/commands/analyze-loading-optimization.mjs
